examples/research_brief/requirements.txt
openai>=2.32.0
httpx>=0.28.1
python-dotenv>=1.2.2
